The api returns the result as file now. Updated home and README

+++ replacer.py
@@ -5,6 +5,7 @@ Created on Thu Apr  4 23:52:05 2019
 @author: Florian
 """
 import os
+from flask import send_file
 
 # make a list of names that should have the ® symbol appended
 brands = open("static/Keyword.txt",'r').read().split('\n')
@@ -13,12 +14,10 @@ brands = open("static/Keyword.txt",'r').read().split('\n')
 def rpl(text):
     for name in brands:
         text = text.replace( name, name + '\u00ae')
-    os.remove("output.txt")
+    if os.path.exists("output.txt"):
+        os.remove("output.txt")
     f = open("output.txt","w+")
     f.write(text)
     f.close()
-    return "OK \n"
-
-    #text = text + '\n'
-    #return text
-
+    return send_file("output.txt", attachment_filename='modified.txt')
+
